// packages/modules/chats/src/lib/chats-types.ts
export type TChatMessage = {
  user_id: string;
  content: string;
  date: string;
  replyIndex?: number;
};

export type TChat = {
  id: string;
  messages: TChatMessage[];
  lastRead: {
    [key: string]: number; // user_id: message_index
  };
  isWriting: {
    [key: string]: boolean; // user_id: boolean
  };
  resolved: boolean;
};